extract backup related code into a dedicated service class
